+++ utils/artnet/__init__.py
@@ -4,7 +4,8 @@
 from .sender import ArtNetSender
 from .dmx_manager import DMXManager, FixtureChannelMap
 from .arbiter import OutputArbiter
+from .live_layer import LiveBuskLayer
 from .shows_artnet_controller import ShowsArtNetController
 
 __all__ = ['ArtNetSender', 'DMXManager', 'FixtureChannelMap',
-           'OutputArbiter', 'ShowsArtNetController']
+           'OutputArbiter', 'LiveBuskLayer', 'ShowsArtNetController']